fix(oauth2) clear authenticated oauth2 headers with multi-auth
### Summary

It was reported that when Kong OAuth 2.0 plugin is configured together
with some other authentication plugin with `conf.anonymous` (logical OR),
the OAuth 2.0 plugin does not clear `X-Authenticated-UserId` and
`X-Authenticated-Scope` headers that it normally only sets on successful
authentication (aka when plugin runs).

This can lead to potential issue on upstream if upstream rely on these
headers and trust that they came from OAuth 2.0 plugin. This change makes
OAuth 2.0 plugin to clear such headers in logical OR scenario.

It is to be noted that Kong itself worked as expected, it is just about
the expectations that upstream service may have made. It is probably
harmless to remove these headers when OAuth 2.0 plugin is configured
in logical OR.