fix(db) when auto-dereferencing fails set value to nil
### Summary

When auto-dereferencing secrets fail, we have two options:

1. keep the value (which means the value is actually a reference such as: `{vault://env/cert-1/key}`
2. set value to `nil`

In both cases the error is also logged.

Original implementation followed 1. but this commit changes it to 2.

Reason being that reference strings can leak to secrets, which they are not meant to.

For example session plugin has secret. If you set secret to `{vault://env/session-secret}`
and the dereferencing fails, the secret becomes `{vault://env/session-secret}`. This
can lead to potential leak of secret on a system that does not resolve secrets correctly.
Or at least it is not good idea that references can become secrets. This commit
changes it so that on failure (we log the warning) and also set the value to `nil`.