feat(hmac-auth) add support for RSA signatures
The hmac-auth plugin allow authentication with HMAC signatures based on
the draft-cavage-http-signatures draft. This commit aims to add support
for RSA signatures as described in the draft, providing a stronger layer
of security via asymmetric encryption.

This implementation has been made with backward compatibility in mind
and only one new field has been added to the DAOs to store the RSA
public key. Depending on the algorithm used during the request, the
plugin will use either the HMAC secret or the RSA public key to verify
the signature.