perf(clustering) conditional rebuilding of router, plugins iterator and balancer on dp (#8519)
Implements conditional rebuilding of `router`, `plugins iterator` and `balancer` on
data planes. This means that DPs will not rebuild router if there were no changes in
routes or services. Similarly, the plugins iterator will not be rebuild if there were
no changes to plugins, and finally balancer in not reinitialized if there are no
changes to upstreams or targets.