fix(rbac) Fix wildcard workspace /kong rbac rule [FTI-3185]
If a customer had a `workspace=* endpoing=/kong` RBAC rule it was not
being enforced. It worked if they specified a workspace, but not all
workspaces. The issue was that the /kong endpoint was a redirect to the
root endpoint, but did not use the typical endpoing naming convention.

Also fixed an issue where not having a top level `endpoint=*` rule
prevented a customer from adding `workspace=current_workspace` rules.
They should still be prevented from adding `workspace=*` rules.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>