chore(proxy-cache) do not copy response data to request's ctx
proxy_cache_hit contains the response data to be shared with the logging
plugins. Originally it was added to the generic nginx request's ctx,
then later it was changed to use Kong's ctx but the previous one was
kept as a backward compatibility. For the next major version this
compatibility will be removed.