feat(conf) allow user-specified OpenResty install path (#8412)
Kong detects and decides which nginx binary to use by searching a couple
hard-coded paths:

* /usr/local/openresty/nginx/sbin
* /opt/openresty/nginx/sbin

...before finally exec-ing out to check PATH for a found nginx
executable. Therefore, if OpenResty is installed at one of the
hard-coded paths, it is always used, leaving the operator no means of
specifying any alternative. In most workflows, this is not a problem at
all, but during local development and testing it can be limiting.

Changing the behavior to search PATH first is one possible way of
addressing this, but it is backwards-incompatible in ways that will
likely cause headaches for operators, developers, package maintainers,
and CI pipeline owners. Therefore, this development
quality-of-life-inspired change introduces a new config option
(`openresty_path`) that short-circuits the aforementioned behavior.
When this value is set, Kong uses it exclusively for locating the nginx
binary.

Using `openresty_path` instead of something like `nginx_bin` allows us
to expand its usage later on (i.e. potentially having a cascading effect
on LUA_PATH/LUA_CPATH).