fix(vaults) do not leak resolved vault references to .kong_env file
### Summary

When Kong prepares a `prefix` directory, it also stores current environment
related to Kong in file called `.kong_env`. As Kong resolves the Vault
references when it starts, the resolved values got leaked to `.kong_env`
file. This was partly because for `vaults-beta` we didn't yet implement
secret rotation, and we decided to also not keep the references around
when they were resolved. Not that we have added the `"$refs"` property
to `kong.configuration`, we can replace the values of configuration with
the references before we write the `.kong_env` file.

This commit fixes that.