feat(vaults) store dao references in $refs (needed for rotation)
### Summary

When there are references used in dao fields with `referenceable=true`,
Kong replaces the references with values when the data is read
(excluding admin api and control planes). When Kong replaces the reference,
it is basically lost, and thus the automatic secret rotation cannot be
implemented. This commit stores the references on returned entities to
`"$refs"` property:

```
local certificate = kong.db.certificates:select(...)
-- the possible reference can be found here:
print(certificate["$refs"].key)
```

There will be helper functions so `"$refs"` property is not intended to
end users.