hotfix(conf) make lua_ssl_trusted_certificate fail softly on launch
When `system` was opt-in, it made sense to "fail noisily" if the current
system didn't have certs in any of the usual places and the user had
opted in to system.

Now that we look for them by default, the noisy behavior is incorrect,
because it makes kong unable to start in places with no system certs.
Instead, the abstence of default certificates gets logged,
but Kong can continue without them.

Related with #8602