fix(plugin/rate-limiting) perform atomic redis increments
The increment function of the Redis policy first gathers which relevant
keys actually exist, and then executes all the INCRBY/EXPIRE operations.
This has a window of error between the existence check and the
(optional) EXPIRE operation.

A solution documented in the Redis documentation
(https://redis.io/commands/incr) is to use a Lua operation in the Redis
server, which operates atomically.

Fixes #5763

Co-authored-by: Javier <javier.guerra@konghq.com>