fix(clustering) Handle connection error on version negotiation
If the CP doesn't respond to the version negotiation request, the DP
will assume the "old" protocol.

A failure to establish the websocket connection repeats the negotiation
attempt, as it would probably indicate a control plane rebooting or real
connectivity issues.  Later errors would (first) attempt only reconnect,
as they would likely be transient issue.  But if that reconnection
fails, it would again retry to renegotiate.