chore(zipkin) Add zipkin prefix to timer logs
These run on a timer so they don't get the namespace automatically
appended to the log line which makes it hard to determine which plugin
they come from:

    [error] 1099#0: *436751 [kong] handler.lua:102 reporter flush failed to request: operation timed out, context: ngx.timer, client: 172.17.0.1, server: 0.0.0.0:8443

I've tried looking for a way to solve this generically that would
benefit other plugins like `rate-limiting` but can't see a way to set
the logger or get access to the plugin name when the timer is running.

I've omitted the square brackets from `zipkin` to make it clearer that
it's not coming from the automatic namespace formatting.