feat(pdk) add kong.request.get_start_time() (#8688)
This adds kong.request.get_start_time as a PDK function. Before this change the business logic for computing this value was contained in kong.log.serialize(), so the only stable-ish API for fetching start time was kong.log.serialize().started_at (kind of wasteful/inelegant). Therefore this seems like a worthwhile inclusion to the PDK.