chore(vault-auth) use vault_auth_vaults dao/api instead of vaults dao/api
### Summary

There is now a core feature called `vaults` (previously called `vaults_beta`) that has
a naming conflict with this plugin:

1. `kong.db.vaults` conflict
2. `:8001/vaults` conflict

It was decided to rename this plugin DAO from `vaults` to `vault_auth_vaults`, and to
have and Admin API of `/vault-auth/*` instead of `/vaults/*`. This commit changes the
plugin to use new names. There is related PR in core here:
https://github.com/Kong/kong/pull/8871

The core is responsible to do schema changes (that hopefully work).

As a reminder, we still need:
1. proper CP/DP compatibility changes (before `3.0.0`)
2. add and run migration tests
3. move this plugin to use new vaults (perhaps after `3.0.0`)