fix(plugin) avoid plugins of default workspace are executed across workspaces (#3274)
In multiple workspaces scenario, if requested to the non-default workspace with no plugins, the global plugins of default workspace will be executed in all phases(`certificate`, `preread`, `rewrite`, `header_filter`, `body_filter`, `log`) except access. This is because we are unable to determine which workspace current was before `access` phase in current architecture. To reduce the impact, I reset the `ctx.plugins` value when current workspace has no plugins. In this way, the `header_filter`, `body_filter`, `log` phrases are no longer executed.

FTI-3297