chore(ldap) ignore referrals (continuation references) in search result
The results of the Search operation are returned as zero or more
SearchResultEntry and/or SearchResultReference messages, followed by a
single SearchResultDone message:
https://www.rfcreader.com/#rfc4511_line1214

The problem is that our ldap client was completely forgetting that
these SearchResultReference messages exist in such a way that when a
SearchResultReference package came in (protocolOp 19), it would be
unintentionally used to perform a search operation.

Here in this change we're explictly ignoring these protocolOp 19
packets. If we encounter these packets in the packet chain, we'll
explictly skip them. This means that our ldap client is now
intentionally choosing to NOT follow referrals sent by the ldap server.

Also, now the Group DN must be prefixed with `group_name_attribute` AND
suffixed with `group_base_dn`.

Basically, the previous regex was not accepting a root base DN. For
example, considering a case where there are nested "OU's". If a
top-level DN such as "ou=dev,o=company" is specified instead of
"ou=role,ou=groups,ou=dev,o=company", the regex would not find a match.

E.g.:

a regex of:
`group_match = "^cn%=[%w-_+:@%s]+%,dc%=ldap%,dc%=mashape%,dc%=com$"`

a group DN of:
`groupdn = "CN=User1,CN=Users,DC=ldap,DC=mashape,DC=com"`

a base DN of:
`gbase_dn = "dc=ldap,dc=mashape,dc=com"`

Resulted in this returning `nil`:
`string.match(groupdn, group_match)`

This was because the `group_match` regex didn't match the string
`groupdn`. Note that in the `group_match` regex, the string must 1)
*exactly* begin with `group_name_attribute` *AND exactly* end with
`group_base_dn` (no substrings allowed). Note that because `groupdn`
contains the `CN=Users` substring, it doesn't match this condition.

With this change, the string must 2) be *prefixed* with
`group_name_attribute` *AND* be *suffixed* with `group_base_dn`. This
way the above `groupdn` should now match this condition, even though it has the `CN=Users` substring.

So now the regex is NOT written to do 1), but 2) instead. With 2), I
believe we're relaxing the restrictions of the regex. Please bear this
in mind when reviewing.

All of this fixes the limitation outlined in the docs right now:
https://docs.konghq.com/hub/kong-inc/ldap-auth-advanced/#notes

And also fixes this FTI: https://konghq.atlassian.net/browse/FTI-2337