fix(*) remove coroutine in access phase
We introduce this in
https://github.com/Kong/kong/commit/744d3b32e47323301bdbe402f9bf1cb2acbf88e8
to persist the behaviour that kong.response.exit can shortcut current
processing. Though this is also documented in PDK document, our code
doesn't seem to rely on it at all.

Opening this to see what will break and open for comments; if merged
this will be a breaking change as this is user-facing, so it can only go
in 3.0 or 4.0.