refactor(handler) trying to make reconfigure more atomic
### Summary

Originally the `reconfigure` event was `atomic` and it was not yielding.
But this caused huge spikes in latencies when new configurations where
applied.

We started to add yielding in different places that happen in `reconfigure`
event. It may now lead to nasty things like: new router, and old plugins
iterator gets used etc. (router from new configuration and plugin iterator
from old configuration).

I needed to refactor the handler so that it could at least support more
atomic flip of `router` and `plugins iterator`. The balancer is still
problematic as it is shared global state and a new balancer cannot
currently be build behind the scenes cooperatively, and then flip it
atomically together with router and plugins iterator. I asked @locao to
research if that could be doable. I leave that to other PR.