fix(dao): plugin priority collision handling
When plugin priorities collide we cannot guarantee a deterministic
result from the sorting function. There are a couple of plugins
that traditionally have colliding priorities like their EE counterparts.
To ensure determinism this commit adds a collision detection that falls
back to a plugin name based sorting.
Also the function moved to the `kong.tools.utils` module to test it in
isolation.

Signed-off-by: Joshua Schmid <jaiks@posteo.de>