fix(rate-limiting) Restore atomic updates to Redis
This is a partial revert of 83bbfd76b (#8213) which restores atomic
updates to Redis as originally added in 6125577eb (#6150).

It retains the additional integration tests that are broken by this
change and will be fixed in a subsequent commit. It's my interpretation
that the atomic updates were only reverted to satisfy these tests.