feat(pdk) add caching to process secrets function
### Summary

Originally process secrets was meant only to be used on CLI where the cache
may not be available (e.g. when parsing `kong.conf`), but we decided later
that process secrets could also be used on runtime for config secrets.

This commit implements similar caching that is currently enable on config cache
function to process secrets function.