fix(hcv) hashicorp vault to not verify ssl (#3353)
### Summary

The vault may be used very early on, e.g. when Kong is loading configuration,
and that happens with CLI (`resty`). The `resty` or `/bin/kong` doesn't currently
support trusted certificates, so it does not make sense to verify server certificates
at this point. We need to revisit this later with a proper support for trusted certs
on CLI.