feat(rate-limiting) support dynamic path filtering
The current implementation of path based rate limiting only allows a
single path identifier per route. However, there are situations where
rate limiting by path would be helpful on a path that is dynamic, such
as including a UUID in the path.

The changes included support this functionality by allowing the
rate_limiting identifier to be automatically set to the path of the
request. This means you don't need to specify which path to rate limit
as it will be done automatically.

The existing test in `04-access_spec.lua`:

https://github.com/Kong/kong/blob/master/spec/03-plugins/23-rate-limiting/04-access_spec.lua#L396

Shows this change working by simply removing the `path` value set in the
config, the test continues to function as expected as the path gets
automatically pulled in.