fix(workspace) get the workspace from the data of "dao:crud" and "crud" event
This bug does not occur when using the legacy events library,
because when the `post_local` method of the old event library is called,
the corresponding `handler` is executed in the current context,
so the `request_id` and workspace can be obtained.

But the new events library doesn't execute any handler in the current context,
so this bug will appear.

When we raise any event,
we should assume that the corresponding handler is running in a completely different context.

* Legacy evnets library: https://github.com/Kong/lua-resty-worker-events
* New events library: https://github.com/Kong/lua-resty-events