chore(dist) add conditional luasec install step
luasec was removed because FIPS build(BoringSSL) doesn't compile against
it (in https://github.com/Kong/kong-ee/pull/2852), but it turns out
there're still couple of places relys on luasocket+luasec to do SSL.
And as in tests we always install luasec as dev dependency, no tests
are failing.

This PR adds it back as a "optional" step, meaning this step will
silently fail for BoringSSL, but will success for OpenSSL. The end
result is we won't break normal build, but will not have like `pg_ssl`
support for FIPS build unfortunately.