fix(locks) use bundled lua-resty-lock from https://github.com/Kong/lu… (#3436)
* fix(locks) use bundled lua-resty-lock from https://github.com/Kong/lua-resty-lock/pull/1

This temporary patch is to avoid a memory leak issue reported by
Coinbase. It can be reproduced by:
- Create 500 upstreams and turn on healthcheck, but not add any targerts
- Setup a hybrid mode cluster, set `db_cache_ttl` to very low (like
0.1s); or use KIC to send frequent config update
- Observe memory increase and the size of `memo` table in the original
lua-resty-lock implementation to grow indefinitely.

The reason for this is likely because the __gc handler is called when
`ref_obj` is executed. It can also be "solved" by turning setting
`jit.off(ref_obj)`.
There're also alternative approach in our fork. Before discuss on which approach
would be a more proper fix in deep, which we will need to do for 3.0
release and onwards, we choose one that's less likely to run into
the same issue again and to unblock the 2.8.1.2 release.

FTI-4071