fix(dao) response "HEALTHCHECK_OFF" for targets with weight 0
Target with weight 0 will not be added into the healthcheck list,
and the default Health status of targets is HEALTHY. For such
targets, Admin API call /upstreams/<upstream>/health reports Health
status as HEALTHY while the healthcheck is not performed for them.
The appropriate status is HEALTHCHECK_OFF.
Fix for Admin API call to provide response with "HEALTHCHECK_OFF"
for targets with weight 0.