chore(conf) remove warning on experimental `AAAA` in `dns_order` configuration (#9214)
### Summary

We added the feature and warning on 8 Sep 2021 (see #7819), and we have not heard
reports on it (perhaps because it was not turned on by default). Thus I removed
the warning of the feature being experimental, and now it is considered a fully
supported one (though I didn't enable it yet by default).