feature(license) restrict license free mode to env vault FT-3177
With this change, an EE license is required to use the aws, hcv or gcp
vaults.  License checks are done at schema validation and vault
retrieval time to ensure that the non-free vaults can be used before
the license has been loaded (i.e. the license itself could be loaded
using a vault reference).