fix(dao) add recursion over entities that need to be cascade-deleted
In a scenario where `consumer` entities were deleted, the
`oauth2_tokens` entities didn't have their caches invalidated. This is
because the `find_cascade_delete_entities` function was only finding
entities up to a maximum of two levels. So in this case where
`consumer` is referenced by `oauth2_credentials` and
`oauth2_credentials` is referenced by `oauth2_tokens` (consumer ->
oauth2_credentials -> oauth2_tokens), only the `consumer` and
`oauth2_credentials` had their cache invalidated, the `oauth2_tokens`
caches had their cache still valid after a DELETE operation. The
`find_cascade_delete_entities` function of the DAO module needs to be
recursive because of this issue. So this PR is doing that, plus it's
adding unit tests along with integration tests for the `oauth2` plugin.

Fix FTI-2023