Fix: Bug in creating a BatchQueue per config in http-log plugin
Fixes kong/kong#9272

Issue content:
The below is the  function which is used to generate a unique queue-Id per plugin config(`http-log` plugin).
In this `string.format`/`fmt` statement there is only `6` format-specifiers but `8` arguments.
https://github.com/Kong/kong/blob/3379a89e3ad62f148c79931a0e7d19adababb319/kong/plugins/http-log/handler.lua#L125-L135
So, this must be ignoring the last two arguments `queue_size` and `flush_timeout` and it will end up using a queue created with a different config.
This can lead to data loss in some scenarios. e.g. when `queue_size: 1` at one place and lets say `queue_size: 10` in another place. If the former happened to be used first then every other config of `queue_size` will effectively be over written to 1 and the below line will take effect.
https://github.com/Kong/kong/blob/3379a89e3ad62f148c79931a0e7d19adababb319/kong/plugins/http-log/handler.lua#L160-L162