perf(router) dynamic router lru cache size based on number of routes (#9309)
* perf(router) dynamic router lru cache size based on number of routes

### Summary

Implements dynamic router LRU cache size based on number of routes.

- By default we initialize positive router LRU cache of 5000 items.
- If there is more than 5000 routes we re-initialize cache to closes 5000,
  e.g. if there is 5001 routes, we reinitialize cache to 10000 items and
  keep the size 10000 until number of routes reaches either over 10000 or
  less than 5001.
- The maximum size of cache is limited to 100000 items.

This should improve our routing performance with high number of routes
when routes are found in cache. This will increase memory usage though,
but it is expected for users to have beefier systems when number of routes
is high (again we limit this to 100000). If previous LRU of 5000 items
took around 5 MB, the dynamic is expected to take 100 MB at most
(per worker).

* code clean

Co-authored-by: chronolaw <chrono_cpp@me.com>