chore(router) better incremental detection without unnecessary hacks
### Summary

The PR #9242 added incremental rebuilding of router but it contained change
detection logic based on new `expression` field, and then it needed to add
a new dao, change schema, and ended up causing issues on EE.

This PR reverts these hacks and uses simpler change detection algorithm
using the `route.updated_at` but bit differently than what the PR #9242
used, and which caused issues because of updated_at is only 1 sec
precision. It still is, but here we store previous build max
updated_at and use that to detect if route is changed or not.

There is a slight overlap as the previous router rebuild new changes
are considered changes in new router build, but this overlap is just
within 1 second, so it should not be a real world problem. We can
always revert to field by field change detection too on such case,
but I think it is overkill.