perf(router) cherry pick router incremental and LRU cache size changes from OSS
### Summary

Implements dynamic router LRU cache size based on number of routes.

- By default we initialize positive router LRU cache of 5000 items.
- If there is more than 5000 routes we re-initialize cache to closes 5000,
  e.g. if there is 5001 routes, we reinitialize cache to 10000 items and
  keep the size 10000 until number of routes reaches either over 10000 or
  less than 5001.
- The maximum size of cache is limited to 100000 items.

This should improve our routing performance with high number of routes
when routes are found in cache. This will increase memory usage though,
but it is expected for users to have beefier systems when number of routes
is high (again we limit this to 100000). If previous LRU of 5000 items
took around 5 MB, the dynamic is expected to take 100 MB at most
(per worker).

Co-authored-by: chronolaw <chrono_cpp@me.com>

* perf(router/atc) incremental ATC router rebuild

The PR #9242 added incremental rebuilding of router but it contained change
detection logic based on new `expression` field, and then it needed to add
a new dao, change schema, and ended up causing issues on EE.

This PR reverts these hacks and uses simpler change detection algorithm
using the `route.updated_at` but bit differently than what the PR #9242
used, and which caused issues because of updated_at is only 1 sec
precision. It still is, but here we store previous build max
updated_at and use that to detect if route is changed or not.

There is a slight overlap as the previous router rebuild new changes
are considered changes in new router build, but this overlap is just
within 1 second, so it should not be a real world problem. We can
always revert to field by field change detection too on such case,
but I think it is overkill.

Co-authored-by: chronolaw <chrono_cpp@me.com>
Co-authored-by: Datong Sun <datong.sun@konghq.com>