fix(ee/routes) guard router rebuild with mutex to avoid concurrent router rebuilds from causing incremental rebuild races. fixes FT-3291
In FT-3291, we discovered that decK sync causes the router to be in a
non-functional state. We knew this is related to incremental router
rebuild, but it does not explain why the same issue did not happen in
OSS. @bungle did some digging and it seems that by setting `route_validation_strategy`
to `off`, the issue also went away.

After some careful logging and investigations, I found that the issue is
caused by the fact that inside `kong/api/routes/routes.lua`, we make
calls to `core_handler.build_router(uuid())` which is not guarded by a
mutex. If sufficient number of routes are present, with high volume of
Admin API modifying the routes the background timer that rebuilds the
router from proxy path will overlap with the router rebuild triggered by
the Admin API collision detection. With the introduction of yielding,
they might be interleaved and cause undefined behavior.

This commit fixes the issue by:

1. Introducing a flag to detect when such race condition happens
2. Guard the router collision detection router rebuild under the same
mutex as proxy path

According to my test, after the change, there are no races detected by
the flag (previous there were many) and the decK sync no longer causes
any 404 on the proxy path afterwards.