chore(clustering) log when skipping a config push
This inserts a debug log entry when a config push is skipped because
there are no connected clients. The side effect is that we can test for
the correct behavior more accurately in our integration tests.

`control_plane.lua` and `wrpc_control_plane.lua` each use a
distinct log prefix, so I simplified existing log messages as well
(hopefully should make things easier when we refactor this code)