fix(external-plugins): handle failing instance creation
This is a fix for https://github.com/Kong/kong/issues/9301
when a request made the plugin fail, the creation of a new instance
was not handling a failed connection properly.
Upon failure to connect to the socket, it was possible to end up
in a state where the instance existed without an instance id,
causing an infinite loop and high CPU utilization.