fix(rbac): avoid the rbac token to be constantly rehashed (#3782)
In this case, `rbac_token` is the plain text token from the kong-admin-token header. `secret.needs_rehash` will always return `true` so kong is updating again the `user_token` and `user_token_ident` in the DB. The new `user_token_ident` is updated to a new ident based in the encrypted token which doesn't match the initial ident generated from plain text `rbac_token`. After this, kong is not able anymore to associate the token to the user.