Refactor datadog hostname/port env vars
This addresses an issue reported here:
https://github.com/Kong/kong/issues/7954

While the tests set a null value to instead inherit the environment
variables, this does not work in a Kubernetes context. For example, if I
have this config:

```
config:
  host: null
  port: 8125
```

The hostname still resolves to `localhost` once it gets through the
schema.lua defaults.

This approach reverses the order of ENV var evaluation, so that
environment variables are used if set, and config if not.