Backport of 8c55a4548eb55a2a905f67e9267c20825794a6ef to 2.8:
fix(*): do not use stale router data if workers are respawned

@Benny-Git reported on #9090 about issue of stale data picked up for router
in case the worker was restarted.

Some insights:

There are several ways to restart workers:
1. you reload them with e.g. kong reload or similar nginx signal
2. worker crashes
3. worker is killed

On case 1 the `init` phase will be re-executed, on 2 and 3 it will not.

This commit ensures that the router is rebuild on init worker if it is not
current.