fix(*) disable jit on m1 computers
There are known issues with JIT and Apple Silicon processors, possibly
related to Apple's Hardened Runtime Capability and the fact that JIT
compilers modify code in memory.
This is a temporary workaround, until more is known about the issue
and a proper fix is created. Note that disabling JIT has considerable
performance penalty, as LuaJIT will run in interpreted mode only. We
deem such performance impact to be acceptable in principle since
Silicon-based machines are used mostly in development, local
testing mode.