chore(release) Deduplicating grep-kong-version.sh script
Currently this script lives in both the kong and kong-ee repos, but we
only use it in those repos to pass the Kong version to KBT. This also
includes the introduction of the `KONG_TAG` variable present in the
next/3.0.x.x branch. That variable is used to determine if this is
a pre-release branch.

Signed-off-by: Tyler Ball <tyler.ball@konghq.com>