feat(aws-lambda): add requestContext field into aws-gateway-compatible input (#9380)
This PR adds a new field `requestContext` into the lambda input data when the aws-lambda plugin is working in `awsgateway_compatible` mode.

The `requestContext` is part of the input data in lambda proxy integration, according to [this AWS document page](https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format), and it contains many fields containing request related information on AWS API Gateway. 

The aws-lambda plugin in Kong provides an AWS API Gateway compatible mode that can generate an input data similar with 
[the example provided in this doc](https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-input-format), but it does not contain the `requestContext` field. When users are writing their lambda functions with some libraries like [aws-serverless-java-container](https://github.com/awslabs/aws-serverless-java-container), it will check if the input data have the `requestContext` field, and if it does not exist, the request will be regarded as malformed.(For example, see [this code snippet](https://github.com/awslabs/aws-serverless-java-container/blob/98482d485dadaf7212141f162e7e35f4ea6c7b6f/aws-serverless-java-container-core/src/main/java/com/amazonaws/serverless/proxy/internal/servlet/AwsProxyHttpServletRequestReader.java#L47-L49))

To make the input data sound more "compatible", this PR adds the `requestContext` field into the input data. Note that many fields in the formal `requestContext` cannot be fetched by Kong(like IAM-related information or Cognito-related auth info), so we're just providing a **best-effort delivery** here. We cannot make assurance that the `requestContext` provided by Kong has every information that the lambda function needs, and it is decided by the lambda function whether the whole payload containing `requestContext` can work normally or not. So the feature is just a good-to-have one.

This should fix FTI-2620, and #7641, in some cases.