tests(clustering): fix version-checking logic in DP tests
This test was failing due to some quirks in the way Kong EE presents its
version string. The version logic could be updated, but I don't want to
make that kind of a change right now. The test in question needs to
identify the data-plane from the output of `GET
/clustering/data-planes`, and it generates a fresh UUID for each time it
connects, so simply testing based on UUID should be sufficient to select
the correct data-plane instance.