chore(plugins-iterator) keep hash table of global plugins (#3887)
### Summary

On PR #3881 we did this, but we cannot test it because Pongo uses latest Kong
code from master branch (and not from PR branch). Thus, the PR is still failing.
Our theory is that if we merge this separately, and then rebase the #3881,
the #3881 should show that we have fixed the issue `exit-transformer`.

Keeping hashtable of globals may turn useful in other contexts too. We will
follow up a CE PR to make delta smaller between EE and CE.