fix(router): regex_priority for identical pattern
Address incorrect priority assignment when multiple routes are defined
with identical paths and different regex priorities.

What happens is that the priority assigned to a certain route reflects
on all the other routes that have identical regex patterns, resunting
in an edge case that may produce wrong prioritization of certain routes
with a consequent incorrect routing of requests.