chore(docs) remove path from admin_gui_url values in kong.conf.default (#3384)
The path does not appear to work at all, and there is an FTI for it to be added so I believe the documentation here is incorrect. I actually filed a Docs PR (https://github.com/Kong/docs.konghq.com/pull/4034) before but they pointed me to correct it here instead as I guess the docs are auto-generated from this file.