Release: 3.0.1 (#9689)
* docs(CHANGELOG) rename unreleased to 3.0.1

* fix(router/atc): properly handle empty fields on the `Route` object (#9470)

Otherwise, it may generate invalid ATC expressions like `() && (http.path == "/foo")`.

* fix(instrumentation) plugin_header_filter instrumentation was not in correct place (#9472)

* fix(*): do not use stale router data if workers are respawned

### Summary

@Benny-Git reported on #9090 about issue of stale data picked up for router
in case the worker was restarted.

Some insights:

There are several ways to restart workers:
1. you reload them with e.g. kong reload or similar nginx signal
2. worker crashes
3. worker is killed

On case 1 the `init` phase will be re-executed, on 2 and 3 it will not.

This commit ensures that the router is rebuild on init worker if it is not
current.

* fix(router/atc): properly handle expressions building error

Under `traditional_compatiable` mode and in rare cases, there is a possibility that ATC generated from the `Route` object contains invalid regexes. When this happens, we need to properly detect the error and report it in the log.

* fix(cmd): kong prepare to write out .kong_process_secrets (#9478) (#9506)

### Summary

Kong Docker image for example starts Kong by running `kong prepare` and
then executes `nginx -c ... -p ...` as seen in:
https://github.com/Kong/docker-kong/blob/master/alpine/docker-entrypoint.sh

This causes issues with Kong secret management references. By default, Kong
passes secrets to nginx using environment variable when using `kong start`,
but here the `nginx` is started directly without calling `kong start` and
thus the secrets are not available for Kong init. This commit fixes that.

Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>

* fix(core): ensure absolute path is used for event sock (#9591)

Co-authored-by: Michael Martin <3277009+flrgh@users.noreply.github.com>

* fix(plugins/zipkin): correctly parse `ot` baggage headers

Co-authored-by: Mayo <mayocream39@yahoo.co.jp>
Co-authored-by: Wangchong Zhou <fffonion@gmail.com>

* docs(changelog): remove incorrect entries

* refactor(handler): trying to make reconfigure more atomic (#9595)

### Summary

Originally the `reconfigure` event was `atomic` and it was not yielding.
But this caused huge spikes in latencies when new configurations where
applied.

We started to add yielding in different places that happen in `reconfigure`
event. It may now lead to nasty things like: new router, and old plugins
iterator gets used etc. (router from new configuration and plugin iterator
from old configuration).

I needed to refactor the handler so that it could at least support more
atomic flip of `router` and `plugins iterator`. The balancer is still
problematic as it is shared global state and a new balancer cannot
currently be build behind the scenes cooperatively, and then flip it
atomically together with router and plugins iterator. I asked @locao to
research if that could be doable. I leave that to other PR.

Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>

* docs(kong-admin-api.yml) update Admin API definition for 3.0.1

* Release 3.0.1

Signed-off-by: VicYP <31416378@qq.com>
Co-authored-by: Enrique García Cota <kikito@gmail.com>
Co-authored-by: Chrono <chrono_cpp@me.com>
Co-authored-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>
Co-authored-by: Michael Martin <3277009+flrgh@users.noreply.github.com>
Co-authored-by: Jay Pathak <52605767+pathak1515@users.noreply.github.com>
Co-authored-by: Mayo <mayocream39@yahoo.co.jp>
Co-authored-by: Wangchong Zhou <fffonion@gmail.com>
Co-authored-by: Michael Martin <flrgh@protonmail.com>